ESXI Guest Network Adapter Bug Work-A-Round
The issue that affects the network adapter of the guest OS. In my case, Windows Server 2012r2. The network adapter will randomly lose connection to the network. Recycling the NIC (Disabling and re-enabling or restarting the NIC) resolves the connectivity issue. This PowerShell script restarts all NICs if a Ping test to 1.1.1.1 fails. It can be modified to restart a single network or Ping a different destination address.